Move `message::Error` to `error::MessageError`
